+++ src/app/admin/honeymoon/excursions/page.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import ExcursionsTab from '../ExcursionsTab';
+import { useHoneymoonApi } from '../HoneymoonContext';
+
+export default function HoneymoonExcursionsPage() {
+    return <ExcursionsTab api={useHoneymoonApi()} />;
+}